requirements.txt
folium==0.20.0
haversine==2.9.0
numpy==2.4.0
pandas==2.3.3
Pillow==12.0.0
plotly==5.24.1
streamlit==1.37.1
streamlit_folium==0.25.3
